+++ porch/api/porch/doc.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 //go:generate go run k8s.io/code-generator/cmd/deepcopy-gen --input-dirs ./... -O zz_generated.deepcopy --go-header-file ../../hack/boilerplate.go.txt
-//go:generate go run k8s.io/code-generator/cmd/openapi-gen --input-dirs github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 --input-dirs k8s.io/apimachinery/pkg/apis/meta/v1,k8s.io/apimachinery/pkg/runtime,k8s.io/apimachinery/pkg/version --output-package github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/openapi -O zz_generated.openapi --go-header-file ../../hack/boilerplate.go.txt
+//go:generate go run k8s.io/code-generator/cmd/openapi-gen --input-dirs github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 --input-dirs k8s.io/apimachinery/pkg/apis/meta/v1,k8s.io/apimachinery/pkg/runtime,k8s.io/apimachinery/pkg/version --output-package github.com/GoogleContainerTools/kpt/porch/api/generated/openapi -O zz_generated.openapi --go-header-file ../../hack/boilerplate.go.txt
 
 // +k8s:deepcopy-gen=package,register
 // +groupName=porch.kpt.dev
