+++ porch/api/porch/v1alpha1/doc.go
@@ -14,9 +14,9 @@
 
 //go:generate go run k8s.io/code-generator/cmd/deepcopy-gen --input-dirs ./ -O zz_generated.deepcopy --go-header-file ../../../hack/boilerplate.go.txt
 //go:generate go run k8s.io/code-generator/cmd/defaulter-gen --input-dirs ./ -O zz_generated.defaults --go-header-file ../../../hack/boilerplate.go.txt
-//go:generate go run k8s.io/code-generator/cmd/client-gen --clientset-name versioned --input-base "" --input github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 --output-package github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/clientset --plural-exceptions PackageRevisionResources:PackageRevisionResources --go-header-file ../../../hack/boilerplate.go.txt
-//go:generate go run k8s.io/code-generator/cmd/lister-gen --input-dirs github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 --output-package github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/listers --go-header-file ../../../hack/boilerplate.go.txt
-//go:generate go run k8s.io/code-generator/cmd/informer-gen --input-dirs github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 --versioned-clientset-package github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/clientset/versioned --listers-package github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/listers --output-package github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/informers --plural-exceptions PackageRevisionResources:PackageRevisionResources --go-header-file ../../../hack/boilerplate.go.txt
+//go:generate go run k8s.io/code-generator/cmd/client-gen --clientset-name versioned --input-base "" --input github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 --output-package github.com/GoogleContainerTools/kpt/porch/api/generated/clientset --plural-exceptions PackageRevisionResources:PackageRevisionResources --go-header-file ../../../hack/boilerplate.go.txt
+//go:generate go run k8s.io/code-generator/cmd/lister-gen --input-dirs github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 --output-package github.com/GoogleContainerTools/kpt/porch/api/generated/listers --go-header-file ../../../hack/boilerplate.go.txt
+//go:generate go run k8s.io/code-generator/cmd/informer-gen --input-dirs github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 --versioned-clientset-package github.com/GoogleContainerTools/kpt/porch/api/generated/clientset/versioned --listers-package github.com/GoogleContainerTools/kpt/porch/api/generated/listers --output-package github.com/GoogleContainerTools/kpt/porch/api/generated/informers --plural-exceptions PackageRevisionResources:PackageRevisionResources --go-header-file ../../../hack/boilerplate.go.txt
 //go:generate go run k8s.io/code-generator/cmd/conversion-gen --input-dirs github.com/GoogleContainerTools/kpt/porch/api/porch,github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1 -O zz_generated.conversion --go-header-file ../../../hack/boilerplate.go.txt
 
 // Api versions allow the api contract for a resource to be changed while keeping
