+++ porch/apiserver/pkg/cmd/server/start.go
@@ -24,11 +24,11 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 
+	clientset "github.com/GoogleContainerTools/kpt/porch/api/generated/clientset/versioned"
+	informers "github.com/GoogleContainerTools/kpt/porch/api/generated/informers/externalversions"
+	sampleopenapi "github.com/GoogleContainerTools/kpt/porch/api/generated/openapi"
 	porchv1alpha1 "github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1"
 	"github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/apiserver"
-	clientset "github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/clientset/versioned"
-	informers "github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/informers/externalversions"
-	sampleopenapi "github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/openapi"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/apiserver/pkg/admission"
