+++ porch/apiserver/pkg/e2e/suite.go
@@ -29,9 +29,9 @@ import (
 	internalpkg "github.com/GoogleContainerTools/kpt/internal/pkg"
 	kptfilev1 "github.com/GoogleContainerTools/kpt/pkg/api/kptfile/v1"
 	porchtest "github.com/GoogleContainerTools/kpt/pkg/test/porch"
+	porchclient "github.com/GoogleContainerTools/kpt/porch/api/generated/clientset/versioned"
 	porchapi "github.com/GoogleContainerTools/kpt/porch/api/porch/v1alpha1"
 	configapi "github.com/GoogleContainerTools/kpt/porch/api/porchconfig/v1alpha1"
-	porchclient "github.com/GoogleContainerTools/kpt/porch/apiserver/pkg/generated/clientset/versioned"
 	"github.com/GoogleContainerTools/kpt/porch/repository/pkg/git"
 	gogit "github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
