+++ porch/go.mod
@@ -2,9 +2,10 @@ module github.com/GoogleContainerTools/kpt/porch
 
 go 1.17
 
-replace github.com/GoogleContainerTools/kpt => ../
-
-replace github.com/GoogleContainerTools/kpt/porch/api => ./api
+replace (
+	github.com/GoogleContainerTools/kpt => ../
+	github.com/GoogleContainerTools/kpt/porch/api => ./api
+)
 
 require (
 	cloud.google.com/go/container v1.2.0
@@ -39,7 +40,6 @@ require (
 	k8s.io/component-base v0.23.5
 	k8s.io/klog/v2 v2.40.1
 	k8s.io/kube-aggregator v0.23.5
-	k8s.io/kube-openapi v0.0.0-20211115234752-e816edb12b65
 	k8s.io/utils v0.0.0-20211208161948-7d6a63dca704
 	sigs.k8s.io/controller-runtime v0.11.0
 	sigs.k8s.io/kustomize/kyaml v0.13.3
@@ -166,6 +166,7 @@ require (
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 	k8s.io/apiextensions-apiserver v0.23.5 // indirect
 	k8s.io/cli-runtime v0.23.5 // indirect
+	k8s.io/kube-openapi v0.0.0-20211115234752-e816edb12b65 // indirect
 	k8s.io/kubectl v0.23.5 // indirect
 	sigs.k8s.io/apiserver-network-proxy/konnectivity-client v0.0.30 // indirect
 	sigs.k8s.io/cli-utils v0.29.2 // indirect
